# perf: remove an extra call to cgo