Before:

```
goos: darwin
goarch: arm64
pkg: github.com/dunglas/frankenphp
BenchmarkHelloWorld
BenchmarkHelloWorld-8   	    3128	    380514 ns/op	   20509 B/op	      33 allocs/op
PASS
ok  	github.com/dunglas/frankenphp	2.284s
```

After:

```
goos: darwin
goarch: arm64
pkg: github.com/dunglas/frankenphp
BenchmarkHelloWorld
BenchmarkHelloWorld-8   	    3144	    362953 ns/op	   20517 B/op	      33 allocs/op
PASS
ok  	github.com/dunglas/frankenphp	2.163s
```